Permite transações em pontos ou reais com saldo decimal
- RewardTransaction.amount muda de Int para Double
- Diálogo de movimentação tem toggle "Pontos / Reais"
- Ao inserir em reais, o valor é convertido para pontos usando a taxa configurada
- Saldo exibido com até 2 casas decimais (zeros trailing removidos)
- Extrato mostra o equivalente em R$ ao lado de cada movimentação
- formatPoints() e formatSigned() para Double

+++ app/src/main/java/com/joaoepedro/tasks/data/AppModels.kt
@@ -60,7 +60,7 @@ data class RewardTransaction(
     val id: String = UUID.randomUUID().toString(),
     val personId: String,
     val type: TransactionType,
-    val amount: Int,
+    val amount: Double,
     val reason: String,
     val createdAt: LocalDate = LocalDate.now(),
     val reversed: Boolean = false
